requirements.txt
torch==2.6.0
transformers==4.48.3
scikit-learn==1.5.1
opencv-python==4.11.0.86
matplotlib==3.9.2
easyocr==1.7.2
google-generativeai==0.7.2
python-dotenv==1.0.1
opencv-python-headless==4.11.0.86
fastapi==0.114.0
numpy==1.26.4
uvicorn==0.30.6
pillow==10.4.0